auth_service/requirements.txt
Flask==2.0.2
Flask-Mail==0.9.1
mysql-connector-python==8.0.26
bcrypt==3.2.0
WTForms==2.3.3
qrcode[pil]==7.3.1
Pillow==9.3.0
Werkzeug==2.0.2
email-validator==1.1.3
APScheduler==3.9.1
Flask-JWT-Extended==4.4.4